+++ src/__init__.py
@@ -1,5 +1,1 @@
 from src.main import app
-
-
-app.state.llm_class = None
-app.state.llm_instance = None
