+++ src/db/redis_client.py
@@ -4,4 +4,7 @@ from redis import Redis
 import redis
 import os
 
-redis = redis.Redis.from_url(os.getenv("REDIS_URL"))
+# src/db/redis_client.py
+import redis
+redis_client = redis.Redis(host="localhost", port=6379, decode_responses=False)
+
