+++ src/util/CustomRepoReader.py
@@ -1,0 +1,7 @@
+from llama_index.readers.github import GithubRepositoryReader
+
+
+class CustomRepoReader(GithubRepositoryReader):
+
+    def __init__(self):
+        self.__super()
